+++ Tests/SwiftLintFrameworkTests/RulesTests.swift
@@ -102,6 +102,10 @@ class RulesTests: XCTestCase {
         verifyRule(FunctionParameterCountRule.description)
     }
 
+    func testGenericTypeName() {
+        verifyRule(GenericTypeNameRule.description)
+    }
+
     func testImplicitGetter() {
         verifyRule(ImplicitGetterRule.description)
     }
@@ -363,6 +367,7 @@ extension RulesTests {
             // ("testForceUnwrapping", testForceUnwrapping),
             ("testFunctionBodyLength", testFunctionBodyLength),
             ("testFunctionParameterCount", testFunctionParameterCount),
+            ("testGenericTypeName", testGenericTypeName),
             ("testImplicitGetter", testImplicitGetter),
             ("testLargeTuple", testLargeTuple),
             ("testLeadingWhitespace", testLeadingWhitespace),